x/bridge/keeper/grpc_query_cw_20_contract_test.go
package keeper_test

import (
	"testing"

	sdk "github.com/cosmos/cosmos-sdk/types"
	"github.com/stretchr/testify/require"
	"google.golang.org/grpc/codes"
	"google.golang.org/grpc/status"

	keepertest "github.com/LimeChain/mantrachain/testutil/keeper"
	"github.com/LimeChain/mantrachain/testutil/nullify"
	"github.com/LimeChain/mantrachain/x/bridge/types"
)

func TestCw20ContractQuery(t *testing.T) {
	keeper, ctx := keepertest.BridgeKeeper(t)
	wctx := sdk.WrapSDKContext(ctx)
	item := createTestCw20Contract(keeper, ctx)
	for _, tc := range []struct {
		desc     string
		request  *types.QueryGetCw20ContractRequest
		response *types.QueryGetCw20ContractResponse
		err      error
	}{
		{
			desc:     "First",
			request:  &types.QueryGetCw20ContractRequest{},
			response: &types.QueryGetCw20ContractResponse{Cw20Contract: item},
		},
		{
			desc: "InvalidRequest",
			err:  status.Error(codes.InvalidArgument, "invalid request"),
		},
	} {
		t.Run(tc.desc, func(t *testing.T) {
			response, err := keeper.Cw20Contract(wctx, tc.request)
			if tc.err != nil {
				require.ErrorIs(t, err, tc.err)
			} else {
				require.NoError(t, err)
				require.Equal(t,
					nullify.Fill(tc.response),
					nullify.Fill(response),
				)
			}
		})
	}
}

